requirements.txt
pandas==1.3.5
numpy==1.23
tensorflow-gpu==2.10.0
